requirements.txt
numpy
sentence-transformers
faiss-cpu
transformers
torch
langchain
pypdf
langchain-text-splitters
groq
